Support login with session
