+++ Desktop-Toggl.xcodeproj/xcuserdata/nghiatran.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,7 +4,7 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>DesktopNoDependency.xcscheme</key>
+		<key>Desktop-Toggl.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>1</integer>
