+++ DesktopCore/DesktopCore/Source/Models/Header.swift
@@ -1,0 +1,17 @@
+//
+//  Header.swift
+//  TogglCore
+//
+//  Created by Nghia Tran on 10/27/18.
+//  Copyright © 2018 com.nsproxy.proxy. All rights reserved.
+//
+
+import Foundation
+
+public struct Header {
+
+    // MARK: - Variable
+    public let key: String
+    public let value: String
+}
+
