+++ DesktopCore/DesktopCore/Source/Protocols/RouteType.swift
@@ -1,0 +1,15 @@
+//
+//  RouteType.swift
+//  TogglCore
+//
+//  Created by Nghia Tran on 10/27/18.
+//  Copyright © 2018 com.nsproxy.proxy. All rights reserved.
+//
+
+import Foundation
+
+protocol RouteType {
+
+    var baseURL: String { get }
+    var path: String { get}
+}
