+++ DesktopCore/TogglCore.xcodeproj/xcuserdata/nghiatran.xcuserdatad/xcschemes/xcschememanagement.plist
@@ -4,10 +4,23 @@
 <dict>
 	<key>SchemeUserState</key>
 	<dict>
-		<key>DesktopCore.xcscheme</key>
+		<key>TogglCore.xcscheme_^#shared#^_</key>
 		<dict>
 			<key>orderHint</key>
 			<integer>0</integer>
+		</dict>
+	</dict>
+	<key>SuppressBuildableAutocreation</key>
+	<dict>
+		<key>BA99994921761B09007BAA60</key>
+		<dict>
+			<key>primary</key>
+			<true/>
+		</dict>
+		<key>BA99995221761B09007BAA60</key>
+		<dict>
+			<key>primary</key>
+			<true/>
 		</dict>
 	</dict>
 </dict>
