+++ DesktopNoDependency/AppDelegate.swift
@@ -18,6 +18,7 @@ class AppDelegate: NSObject, NSApplicationDelegate {
     // MARK: - Variable
     lazy var app: Application = Application.default()
 
+    // MARK: - Public
     func applicationDidFinishLaunching(_ aNotification: Notification) {
         app.start()
     }
